+++ src/main/java/com/example/entities/Tag.java
@@ -24,7 +24,7 @@ public class Tag {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "tag_id")
-	private int id;
+	private long id;
 	@Column(name="title")
 	private String title;
 	
@@ -37,18 +37,18 @@ public class Tag {
 		super();
 	}
 	
-	public Tag(int id, String title, Set<FlashcardSet> flashcardSet) {
+	public Tag(long id, String title, Set<FlashcardSet> flashcardSet) {
 		super();
 		this.id = id;
 		this.title = title;
 		this.fs = flashcardSet;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
